requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
deepgram-sdk==2.8.0
dropbox==11.36.0
elevenlabs==0.2.7
executing==1.2.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
langdetect==1.0.9
multidict==6.0.4
openai==0.27.4
Pillow==9.5.0
pure-eval==0.2.2
pydantic==1.10.7
pydub==0.25.1
python-dotenv==1.0.0
python-telegram-bot==20.2
pytz==2023.3
requests==2.29.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.10.0
stack-data==0.6.2
tiktoken==0.4.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
yarl==1.9.2
youtube-transcript-api==0.6.1